pyproject.toml
[project]
name = "yapoc"
version = "0.1.0"
description = "Yet Another Python OpenClaw — an autonomous multi-agent system where a hierarchy of specialized AI agents collaborate to plan, build, and self-manage tasks via file-based communication."
authors = [
    {name = "Nikita Zaitsev",email = "partner.Nikita.Zaitsev@philips.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "aiofiles>=24.1.0",
    "watchdog>=6.0.0",
    "loguru>=0.7.0",
    "httpx>=0.28.0",
    "typer>=0.15.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "numpy (>=2.4.3,<3.0.0)",
    "psutil (>=7.2.2,<8.0.0)",
    "sentence-transformers (>=5.4.0,<6.0.0)",
]


[project.scripts]
yapoc = "app.cli.main:app"
yapoc-agent = "app.agents.base.runner_entry:main"

[tool.poetry]
packages = [{include = "app"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)"
]
